refactor: the @blockera/packages architecture
