+++ packages/features/package.json
@@ -21,5 +21,5 @@
 	"engines": {
 		"node": ">=12"
 	},
-	"main": "js/index.js"
+	"main": "src/index.js"
 }
