+++ packages/features/src/Contracts/index.php
@@ -1,0 +1,8 @@
+<?php
+/**
+ * We can't change the world, but we can make a block of it nice.
+ *
+ * Security file.
+ *
+ * @package src/Contracts/index.php
+ */
