+++ packages/features/src/Traits/Singleton.php
